New lookup service format

+++ src/SHIP/SHIPLookupService.ts
@@ -1,6 +1,6 @@
-import { LookupService, LookupQuestion, LookupAnswer, LookupFormula } from '@bsv/overlay'
+import { LookupService, LookupQuestion, LookupAnswer, LookupFormula, AdmissionMode, OutputAdmittedByTopic, OutputSpent, SpendNotificationMode } from '@bsv/overlay'
 import { SHIPStorage } from './SHIPStorage.js'
-import { Script, PushDrop, Utils } from '@bsv/sdk'
+import { PushDrop, Utils } from '@bsv/sdk'
 import { SHIPQuery } from 'src/types.js'
 import SHIPLookupDocs from './SHIPLookup.docs.js'
 
@@ -11,18 +11,15 @@ import SHIPLookupDocs from './SHIPLookup.docs.js'
  * within the overlay network.
  */
 export class SHIPLookupService implements LookupService {
+  admissionMode: AdmissionMode = 'locking-script'
+  spendNotificationMode: SpendNotificationMode = 'none'
   constructor(public storage: SHIPStorage) { }
 
-  /**
-   * Handles the addition of a new output to the topic.
-   * @param txid - The transaction ID containing the output.
-   * @param outputIndex - The index of the output in the transaction.
-   * @param outputScript - The script of the output to be processed.
-   * @param topic - The topic associated with the output.
-   */
-  async outputAdded?(txid: string, outputIndex: number, outputScript: Script, topic: string): Promise<void> {
+  async outputAdmittedByTopic(payload: OutputAdmittedByTopic): Promise<void> {
+    if (payload.mode !== 'locking-script') throw new Error('Invalid payload')
+    const { topic, lockingScript, txid, outputIndex } = payload
     if (topic !== 'tm_ship') return
-    const result = PushDrop.decode(outputScript)
+    const result = PushDrop.decode(lockingScript)
     const shipIdentifier = Utils.toUTF8(result.fields[0])
     const identityKey = Utils.toHex(result.fields[1])
     const domain = Utils.toUTF8(result.fields[2])
@@ -31,33 +28,17 @@ export class SHIPLookupService implements LookupService {
     await this.storage.storeSHIPRecord(txid, outputIndex, identityKey, domain, topicSupported)
   }
 
-  /**
-   * Handles the spending of an output in the topic.
-   * @param txid - The transaction ID of the spent output.
-   * @param outputIndex - The index of the spent output.
-   * @param topic - The topic associated with the spent output.
-   */
-  async outputSpent?(txid: string, outputIndex: number, topic: string): Promise<void> {
+  async outputSpent(payload: OutputSpent): Promise<void> {
+    if (payload.mode !== 'none') throw new Error('Invalid payload')
+    const { topic, txid, outputIndex } = payload
     if (topic !== 'tm_ship') return
     await this.storage.deleteSHIPRecord(txid, outputIndex)
   }
 
-  /**
-   * Handles the deletion of an output in the topic.
-   * @param txid - The transaction ID of the deleted output.
-   * @param outputIndex - The index of the deleted output.
-   * @param topic - The topic associated with the deleted output.
-   */
-  async outputDeleted?(txid: string, outputIndex: number, topic: string): Promise<void> {
-    if (topic !== 'tm_ship') return
+  async outputEvicted (txid: string, outputIndex: number): Promise<void> {
     await this.storage.deleteSHIPRecord(txid, outputIndex)
   }
 
-  /**
-   * Answers a lookup query.
-   * @param question - The lookup question to be answered.
-   * @returns A promise that resolves to a lookup answer or formula.
-   */
   async lookup(question: LookupQuestion): Promise<LookupAnswer | LookupFormula> {
     if (question.query === undefined || question.query === null) {
       throw new Error('A valid query must be provided!')
@@ -84,18 +65,10 @@ export class SHIPLookupService implements LookupService {
     return await this.storage.findRecord({ domain, topics, identityKey })
   }
 
-  /**
-   * Returns documentation specific to this overlay lookup service.
-   * @returns A promise that resolves to the documentation string.
-   */
   async getDocumentation(): Promise<string> {
     return SHIPLookupDocs
   }
 
-  /**
-   * Returns metadata associated with this lookup service.
-   * @returns A promise that resolves to an object containing metadata.
-   */
   async getMetaData(): Promise<{
     name: string
     shortDescription: string
